Begin graphics rework
I need to start with the LCD and derive the right pixel values from VRAM in order to sync this all properly with the clock. The existing approach starts with VRAM and maps it to the LCD once per frame. That only works for static backgrounds and won't handle things changing mid frame, like window drawing or scrollX/Y changes.