Major refactoring of flags
Flags previously used a dual representation where each was stored in both the f register and a dedicated boolean. The lack of a single source of truth was causing subtle bugs. This commit unifies on a more direct approach that updates the f register directly in each instruction implementation. It's more obvious what is going on and more blargg tests pass as a result of clearing away hidden bugs.