Work on POP timing
POP timing is now correct in principle but the mooneye test doesn't pass. I can force it pass by moving the internal delay to the end of the instruction instead of the start, where it should be. That indicates an OAM timing problem somewhere else that has a knock on effect on the POP, and probably PUSH, mooneye tests.